+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Unreleased
 
 .. vendor-insert-here
 
+- Support multi-document GitLab CI configuration files. (:issue:`561`)
 - Update vendored schemas: bitbucket-pipelines, mergify, renovate (2026-08-16)
 
 0.38.0